+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/autoconfigure/aad/AADAuthenticationProperties.java
@@ -18,8 +18,8 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
@@ -150,6 +150,11 @@ public class AADAuthenticationProperties implements InitializingBean {
             return allowedGroupIds;
         }
 
+        /**
+         * Set the allowed group ids.
+         *
+         * @param allowedGroupIds Allowed group ids.
+         */
         public void setAllowedGroupIds(Set<String> allowedGroupIds) {
             this.allowedGroupIds = allowedGroupIds;
         }
@@ -162,6 +167,10 @@ public class AADAuthenticationProperties implements InitializingBean {
             this.allowedGroupNames = allowedGroupNames;
         }
 
+        @Deprecated
+        @DeprecatedConfigurationProperty(
+            reason = "enable-full-list is not easy to understand.",
+            replacement = "allowed-group-ids: all")
         public Boolean getEnableFullList() {
             return enableFullList;
         }
@@ -380,7 +389,7 @@ public class AADAuthenticationProperties implements InitializingBean {
     }
 
     @Override
-    public void afterPropertiesSet() throws Exception {
+    public void afterPropertiesSet() {
 
         if (!StringUtils.hasText(baseUri)) {
             baseUri = "https://login.microsoftonline.com/";
@@ -407,6 +416,14 @@ public class AADAuthenticationProperties implements InitializingBean {
                 + "the prefix of azure.activedirectory.graph-membership-uri. "
                 + "azure.activedirectory.graph-base-uri = " + graphBaseUri + ", "
                 + "azure.activedirectory.graph-membership-uri = " + graphMembershipUri + ".");
+        }
+
+        Set<String> allowedGroupIds = userGroup.getAllowedGroupIds();
+        if (allowedGroupIds.size() > 1 && allowedGroupIds.contains("all")) {
+            throw new IllegalStateException("When azure.activedirectory.user-group.allowed-group-ids contains 'all', "
+                + "no other group ids can be configured. "
+                + "But actually azure.activedirectory.user-group.allowed-group-ids="
+                + allowedGroupIds);
         }
 
         if (!StringUtils.hasText(tenantId)) {